feat(anna): Added a working CLI (#1258)
With this change, the water flows through the pipes, with Gets, Sets &
Deletes primarily working. Some minor issues remain, but I'll do those
separately.

Next steps:
https://github.com/hydro-project/hydroflow/issues/1253
https://github.com/hydro-project/hydroflow/issues/1257
https://github.com/hydro-project/hydroflow/issues/1256
https://github.com/hydro-project/hydroflow/issues/1255
https://github.com/hydro-project/hydroflow/issues/1254

+++ datastores/gossip_kv/server/model.rs
@@ -22,7 +22,7 @@ pub type TableMap<C> = MapUnionHashMap<TableName, Table<C>>;
 
 /// Timestamps used in the model.
 // TODO: This will be updated to use a more sophisticated clock type with https://github.com/hydro-project/hydroflow/issues/1207.
-pub type Clock = Max<usize>;
+pub type Clock = Max<u64>;
 
 /// TableMap element to upsert a row in an existing TableMap.
 ///
